pyproject.toml
[tool.poetry]
name = "twitter_search_client"
version = "0.1.0"
description = ""
authors = ["lestherll <ljllacuna5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
types-requests = "^2.27.30"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true